pipeline refactoring: getters and expressions
