+++ core/services/pipeline/task_params.go
@@ -3,7 +3,6 @@ package pipeline
 import (
 	"encoding/hex"
 	"encoding/json"
-	"fmt"
 	"math"
 	"math/big"
 	"net/url"
@@ -46,156 +45,6 @@ func ResolveParam(out PipelineParamUnmarshaler, getters []GetterFunc) error {
 		return err
 	}
 	return nil
-}
-
-type GetterFunc func() (interface{}, error)
-
-func From(getters ...interface{}) []GetterFunc {
-	var gfs []GetterFunc
-	for _, g := range getters {
-		switch v := g.(type) {
-		case GetterFunc:
-			gfs = append(gfs, v)
-
-		default:
-			// If a bare value is passed in, create a simple getter
-			gfs = append(gfs, func() (interface{}, error) {
-				return v, nil
-			})
-		}
-	}
-	return gfs
-}
-
-func VarExpr(s string, vars Vars) GetterFunc {
-	return func() (interface{}, error) {
-		trimmed := strings.TrimSpace(s)
-		if len(trimmed) == 0 {
-			return nil, ErrParameterEmpty
-		}
-		isVariableExpr := strings.Count(trimmed, "$") == 1 && trimmed[:2] == "$(" && trimmed[len(trimmed)-1] == ')'
-		if !isVariableExpr {
-			return nil, ErrParameterEmpty
-		}
-		keypath := strings.TrimSpace(trimmed[2 : len(trimmed)-1])
-		val, err := vars.Get(keypath)
-		if err != nil {
-			return nil, err
-		} else if as, is := val.(error); is {
-			return nil, errors.Wrapf(ErrTooManyErrors, "VarExpr: %v", as)
-		}
-		return val, nil
-	}
-}
-
-func JSONWithVarExprs(s string, vars Vars, allowErrors bool) GetterFunc {
-	return func() (interface{}, error) {
-		if strings.TrimSpace(s) == "" {
-			return nil, ErrParameterEmpty
-		}
-		replaced := variableRegexp.ReplaceAllFunc([]byte(s), func(expr []byte) []byte {
-			keypathStr := strings.TrimSpace(string(expr[2 : len(expr)-1]))
-			return []byte(fmt.Sprintf(`{ "__chainlink_var_expr__": "%v" }`, keypathStr))
-		})
-		var val interface{}
-		err := json.Unmarshal(replaced, &val)
-		if err != nil {
-			return nil, errors.Wrapf(ErrBadInput, "while interpolating variables in JSON payload: %v; got input: %v", err, replaced)
-		}
-		return mapGoValue(val, func(val interface{}) (interface{}, error) {
-			if m, is := val.(map[string]interface{}); is {
-				maybeKeypath, exists := m["__chainlink_var_expr__"]
-				if !exists {
-					return val, nil
-				}
-				keypath, is := maybeKeypath.(string)
-				if !is {
-					return nil, errors.New("you cannot use __chainlink_var_expr__ in your JSON")
-				}
-				newVal, err := vars.Get(keypath)
-				if err != nil {
-					return nil, err
-				} else if err, is := newVal.(error); is && !allowErrors {
-					return nil, errors.Wrapf(ErrBadInput, "JSONWithVarExprs: %v", err)
-				}
-				return newVal, nil
-			}
-			return val, nil
-		})
-	}
-}
-
-func mapGoValue(v interface{}, fn func(val interface{}) (interface{}, error)) (x interface{}, err error) {
-	type item struct {
-		val         interface{}
-		parentMap   map[string]interface{}
-		parentKey   string
-		parentSlice []interface{}
-		parentIdx   int
-	}
-
-	stack := []item{{val: v}}
-	var current item
-
-	for len(stack) > 0 {
-		current = stack[0]
-		stack = stack[1:]
-
-		val, err := fn(current.val)
-		if err != nil {
-			return nil, err
-		}
-
-		if current.parentMap != nil {
-			current.parentMap[current.parentKey] = val
-		} else if current.parentSlice != nil {
-			current.parentSlice[current.parentIdx] = val
-		}
-
-		if asMap, isMap := val.(map[string]interface{}); isMap {
-			for key := range asMap {
-				stack = append(stack, item{val: asMap[key], parentMap: asMap, parentKey: key})
-			}
-		} else if asSlice, isSlice := val.([]interface{}); isSlice {
-			for i := range asSlice {
-				stack = append(stack, item{val: asSlice[i], parentSlice: asSlice, parentIdx: i})
-			}
-		}
-	}
-	return v, nil
-}
-
-func NonemptyString(s string) GetterFunc {
-	return func() (interface{}, error) {
-		trimmed := strings.TrimSpace(s)
-		if len(trimmed) == 0 {
-			return nil, ErrParameterEmpty
-		}
-		return trimmed, nil
-	}
-}
-
-func Input(inputs []Result, index int) GetterFunc {
-	return func() (interface{}, error) {
-		if len(inputs)-1 < index {
-			return nil, ErrParameterEmpty
-		}
-		return inputs[index].Value, inputs[index].Error
-	}
-}
-
-func Inputs(inputs []Result) GetterFunc {
-	return func() (interface{}, error) {
-		var vals []interface{}
-		for _, input := range inputs {
-			if input.Error != nil {
-				vals = append(vals, input.Error)
-			} else {
-				vals = append(vals, input.Value)
-			}
-		}
-		return vals, nil
-	}
 }
 
 type StringParam string
