+++ core/services/pipeline/task_params_test.go
@@ -430,82 +430,6 @@ func TestJSONPathParam_UnmarshalPipelineParam(t *testing.T) {
 	}
 }
 
-func TestVarExpr(t *testing.T) {
-	t.Parallel()
-
-	vars := createTestVars()
-
-	tests := []struct {
-		expr   string
-		result interface{}
-		err    error
-	}{
-		// no errors
-		{" $(  foo.bar  ) ", "value", nil},
-		{" $(  zet)", 123, nil},
-		{"$(arr.1  ) ", 200, nil},
-		// errors
-		{" $(  missing)", nil, pipeline.ErrKeypathNotFound},
-		{" $$(  zet)", nil, pipeline.ErrParameterEmpty},
-		{"  ", nil, pipeline.ErrParameterEmpty},
-	}
-
-	for _, test := range tests {
-		test := test
-		t.Run(test.expr, func(t *testing.T) {
-			t.Parallel()
-
-			getter := pipeline.VarExpr(test.expr, vars)
-			v, err := getter()
-			if test.err == nil {
-				require.NoError(t, err)
-				require.Equal(t, test.result, v)
-			} else {
-				require.ErrorIs(t, err, test.err)
-			}
-		})
-	}
-}
-
-func TestJSONWithVarExprs(t *testing.T) {
-	t.Parallel()
-
-	vars := createTestVars()
-
-	tests := []struct {
-		json   string
-		field  string
-		result interface{}
-		err    error
-	}{
-		// no errors
-		{`{ "x": $(zet) }`, "x", 123, nil},
-		{`{ "x": { "y": $(zet) } }`, "x", map[string]interface{}{"y": 123}, nil},
-		{`{ "z": "foo" }`, "z", "foo", nil},
-		// errors
-		{`{ "x": $(missing) }`, "x", nil, pipeline.ErrKeypathNotFound},
-		{`{ "x": "$(zet)" }`, "x", "$(zet)", pipeline.ErrBadInput},
-		{`{ "$(foo.bar)": $(zet) }`, "value", 123, pipeline.ErrBadInput},
-	}
-
-	for _, test := range tests {
-		test := test
-		t.Run(test.json, func(t *testing.T) {
-			t.Parallel()
-
-			getter := pipeline.JSONWithVarExprs(test.json, vars, false)
-			v, err := getter()
-			if test.err != nil {
-				require.ErrorIs(t, err, test.err)
-			} else {
-				require.NoError(t, err)
-				m := v.(map[string]interface{})
-				require.Equal(t, test.result, m[test.field])
-			}
-		})
-	}
-}
-
 func TestResolveValue(t *testing.T) {
 	t.Parallel()
 
@@ -604,133 +528,5 @@ func TestResolveValue(t *testing.T) {
 		require.Equal(t, expectedErr, err)
 
 		param.AssertExpectations(t)
-	})
-}
-
-func TestGetters_VarExpr(t *testing.T) {
-	t.Parallel()
-
-	vars := pipeline.NewVarsFrom(map[string]interface{}{
-		"foo": map[string]interface{}{
-			"bar": 42,
-		},
-	})
-
-	tests := []struct {
-		expr  string
-		value interface{}
-		err   error
-	}{
-		{"$(foo.bar)", 42, nil},
-		{" $(foo.bar)", 42, nil},
-		{"$(foo.bar) ", 42, nil},
-		{"$( foo.bar)", 42, nil},
-		{"$(foo.bar )", 42, nil},
-		{"$( foo.bar )", 42, nil},
-		{" $( foo.bar )", 42, nil},
-		{"$()", nil, pipeline.ErrVarsRoot},
-		{"$(foo.bar", nil, pipeline.ErrParameterEmpty},
-		{"$foo.bar)", nil, pipeline.ErrParameterEmpty},
-		{"(foo.bar)", nil, pipeline.ErrParameterEmpty},
-		{"foo.bar", nil, pipeline.ErrParameterEmpty},
-	}
-
-	for _, test := range tests {
-		test := test
-		t.Run(test.expr, func(t *testing.T) {
-			t.Parallel()
-
-			val, err := pipeline.VarExpr(test.expr, vars)()
-			require.Equal(t, test.value, val)
-			require.Equal(t, test.err, errors.Cause(err))
-		})
-	}
-}
-
-func TestGetters_NonemptyString(t *testing.T) {
-	t.Parallel()
-
-	t.Run("returns any non-empty string", func(t *testing.T) {
-		t.Parallel()
-
-		val, err := pipeline.NonemptyString("foo bar")()
-		require.NoError(t, err)
-		require.Equal(t, "foo bar", val)
-	})
-
-	t.Run("returns ErrParameterEmpty when given an empty string (including only spaces)", func(t *testing.T) {
-		t.Parallel()
-
-		_, err := pipeline.NonemptyString("")()
-		require.Equal(t, pipeline.ErrParameterEmpty, errors.Cause(err))
-		_, err = pipeline.NonemptyString(" ")()
-		require.Equal(t, pipeline.ErrParameterEmpty, errors.Cause(err))
-	})
-}
-
-func TestGetters_Input(t *testing.T) {
-	t.Parallel()
-
-	t.Run("returns the requested input's Value and Error if they exist", func(t *testing.T) {
-		t.Parallel()
-
-		expectedVal := "bar"
-		expectedErr := errors.New("some err")
-		val, err := pipeline.Input([]pipeline.Result{{Value: "foo"}, {Value: expectedVal, Error: expectedErr}, {Value: "baz"}}, 1)()
-		require.Equal(t, expectedVal, val)
-		require.Equal(t, expectedErr, err)
-	})
-
-	t.Run("returns ErrParameterEmpty if the specified input does not exist", func(t *testing.T) {
-		t.Parallel()
-
-		_, err := pipeline.Input([]pipeline.Result{{Value: "foo"}}, 1)()
-		require.Equal(t, pipeline.ErrParameterEmpty, errors.Cause(err))
-	})
-}
-
-func TestGetters_Inputs(t *testing.T) {
-	t.Parallel()
-
-	theErr := errors.New("some issue")
-
-	tests := []struct {
-		name        string
-		inputs      []pipeline.Result
-		expected    []interface{}
-		expectedErr error
-	}{
-		{
-			"returns the values and errors",
-			[]pipeline.Result{
-				{Value: "foo"},
-				{Error: theErr},
-				{Value: "baz"},
-			},
-			[]interface{}{"foo", theErr, "baz"}, nil,
-		},
-	}
-
-	for _, test := range tests {
-		test := test
-		t.Run("returns all of the inputs' Values if the provided inputs meet the spec", func(t *testing.T) {
-			t.Parallel()
-
-			val, err := pipeline.Inputs(test.inputs)()
-			require.Equal(t, test.expectedErr, errors.Cause(err))
-			require.Equal(t, test.expected, val)
-		})
-	}
-}
-
-func createTestVars() pipeline.Vars {
-	return pipeline.NewVarsFrom(map[string]interface{}{
-		"foo": map[string]interface{}{
-			"bar": "value",
-		},
-		"zet": 123,
-		"arr": []interface{}{
-			100, 200, 300,
-		},
 	})
 }
